--- chart_pattern/code/main.py
@@ -1,6 +1,0 @@
-import flag_pattern, pennant_pattern, cup_handle_pattern, gap_pattern
-
-# flag_pattern.start()
-# pennant_pattern.start()
-cup_handle_pattern.start()
-# gap_pattern.start()
