--- chart_pattern/workstages.txt
@@ -1,3 +1,0 @@
-python -m venv venv
-
-pip install pandas mplfinance tqdm
