--- community/docker-compose.yml
@@ -1,38 +1,0 @@
-services:
-  db:
-    image: postgres:15
-    environment:
-      POSTGRES_USER: user
-      POSTGRES_PASSWORD: password
-      POSTGRES_DB: stockmind
-    ports:
-      - "5432:5432"
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-      - ./init_db.sql:/docker-entrypoint-initdb.d/init_db.sql
-    restart: unless-stopped
-    networks:
-      - stockmind-network
-
-  crawler:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    depends_on:
-      - db
-    volumes:
-      - ./data:/app/data
-      - ./logs:/app/logs
-      - ./src:/app/src
-    environment:
-      - PYTHONPATH=/app/src
-    command: python /app/src/migrate_csv_to_db.py
-    networks:
-      - stockmind-network
-
-volumes:
-  postgres_data:
-
-networks:
-  stockmind-network:
-    driver: bridge
