--- community/ts-crawler/.env.example
@@ -1,9 +1,0 @@
-# PostgreSQL Database Configuration
-DB_HOST=localhost
-DB_PORT=5432
-DB_NAME=stockmind
-DB_USER=user
-DB_PASSWORD=password
-
-# Scraper Configuration
-USER_AGENT=Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36
