--- community/ts-crawler/.gitignore
@@ -1,5 +1,0 @@
-node_modules/
-dist/
-.env
-*.log
-.DS_Store
