--- community/ts-crawler/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "yahoo-community-scraper",
-  "version": "1.0.0",
-  "description": "TypeScript-based Yahoo Finance community crawler with HTML parsing",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "ts-node src/index.ts"
-  },
-  "keywords": ["yahoo-finance", "scraper", "community", "stock"],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "axios": "^1.6.5",
-    "cheerio": "^1.0.0-rc.12",
-    "pg": "^8.11.3",
-    "csv-writer": "^1.6.0",
-    "dotenv": "^16.3.1"
-  },
-  "devDependencies": {
-    "@types/node": "^20.11.5",
-    "@types/pg": "^8.10.9",
-    "typescript": "^5.3.3",
-    "ts-node": "^10.9.2"
-  }
-}
