--- community/ts-crawler/test-parser.js
@@ -1,140 +1,0 @@
-/**
- * Test script to verify parsing logic works correctly
- * Uses sample HTML string (not fetched from Yahoo)
- */
-
-const { YahooCommunityScraper } = require('./dist/YahooCommunityScraper');
-
-// Sample HTML with key data-testid selectors (simplified)
-const sampleHtml = `
-<!DOCTYPE html>
-<html>
-<head><title>AAPL Stock Quote</title></head>
-<body>
-  <script>
-    window.YAHOO = {
-      context: {
-        dispatcher: {
-          stores: {
-            QuoteSummaryStore: {
-              symbol: "AAPL",
-              price: {
-                shortName: "Apple Inc.",
-                regularMarketPrice: { raw: 204.08 }
-              },
-              marketState: "REGULAR",
-              currency: "USD"
-            }
-          }
-        }
-      }
-    };
-  </script>
-
-  <div data-testid="quote-hdr">
-    <div class="symbol">AAPL</div>
-  </div>
-
-  <section data-testid="quote-title">
-    <h1>Apple Inc. (AAPL)</h1>
-  </section>
-
-  <span data-testid="qsp-price">204.08</span>
-
-  <div class="exchange">NasdaqGS - NasdaqGS Real Time Price • USD</div>
-
-  <div class="spot-im-container" data-spotid="sp_R2cfqc5q"></div>
-
-  <script>
-    var spotConfig = {
-      "spotId": "sp_R2cfqc5q",
-      "article": "AAPL"
-    };
-  </script>
-</body>
-</html>
-`;
-
-console.log('🧪 Testing YahooCommunityScraper parsing logic...\n');
-console.log('='.repeat(60));
-
-try {
-  const scraper = new YahooCommunityScraper();
-
-  // Test 1: Parse static HTML
-  console.log('\n📊 Test 1: Parse Stock Header');
-  console.log('-'.repeat(60));
-  const stockHeader = scraper.parseStaticHtml(sampleHtml);
-  console.log('Result:');
-  console.table(stockHeader);
-
-  // Verify results
-  const checks = [
-    { name: 'Symbol extracted', pass: stockHeader.symbol === 'AAPL' },
-    { name: 'Name extracted', pass: stockHeader.name.includes('Apple') },
-    { name: 'Price is number', pass: typeof stockHeader.price === 'number' },
-    { name: 'Price is correct', pass: stockHeader.price === 204.08 },
-    { name: 'Currency extracted', pass: stockHeader.currency === 'USD' },
-    { name: 'Exchange extracted', pass: stockHeader.exchange.includes('Nasdaq') }
-  ];
-
-  console.log('\n✅ Validation Checks:');
-  console.log('-'.repeat(60));
-  checks.forEach(check => {
-    console.log(`${check.pass ? '✅' : '❌'} ${check.name}`);
-  });
-
-  const allPassed = checks.every(c => c.pass);
-
-  console.log('\n' + '='.repeat(60));
-  if (allPassed) {
-    console.log('🎉 All tests PASSED! Parsing logic works correctly.');
-  } else {
-    console.log('⚠️  Some tests failed. Check implementation.');
-  }
-  console.log('='.repeat(60));
-
-  // Test 2: Spot ID extraction (using private method via parsing HTML)
-  console.log('\n📊 Test 2: Spot ID Extraction');
-  console.log('-'.repeat(60));
-  const spotIdMatch = sampleHtml.match(/"spotId"\s*:\s*"([a-zA-Z0-9_]+)"/);
-  if (spotIdMatch && spotIdMatch[1]) {
-    console.log(`✅ Found Spot ID: ${spotIdMatch[1]}`);
-  } else {
-    console.log('❌ Spot ID not found');
-  }
-
-  // Test 3: Content sanitization
-  console.log('\n📊 Test 3: Content Sanitization');
-  console.log('-'.repeat(60));
-  const testContent = '<p>This is a <strong>test</strong> comment</p> &nbsp;&amp;&lt;script&gt;';
-  const sanitized = testContent
-    .replace(/<[^>]*>?/gm, '')
-    .replace(/&nbsp;/g, ' ')
-    .replace(/&amp;/g, '&')
-    .replace(/&lt;/g, '<')
-    .replace(/&gt;/g, '>')
-    .replace(/\s+/g, ' ')
-    .trim();
-
-  console.log('Input:    ', testContent);
-  console.log('Sanitized:', sanitized);
-  console.log('✅ HTML tags removed and entities handled');
-
-  console.log('\n' + '='.repeat(60));
-  console.log('📝 Conclusion:');
-  console.log('-'.repeat(60));
-  console.log('✅ TypeScript compilation successful');
-  console.log('✅ HTML parsing logic works correctly');
-  console.log('✅ Price extraction handles numbers properly');
-  console.log('✅ Spot ID regex pattern works');
-  console.log('✅ Content sanitization works');
-  console.log('\n⚠️  Note: Yahoo Finance is rate-limiting actual HTTP requests.');
-  console.log('   This is expected and not a bug in the implementation.');
-  console.log('   The crawler logic is correct and production-ready.');
-  console.log('='.repeat(60) + '\n');
-
-} catch (error) {
-  console.error('❌ Error:', error);
-  process.exit(1);
-}
