--- mlruns/0/meta.yaml
@@ -1,6 +1,0 @@
-artifact_location: mlflow-artifacts:/0
-creation_time: 1756105264987
-experiment_id: '0'
-last_update_time: 1756105264987
-lifecycle_stage: active
-name: Default
