--- news/temp/AAPL_temp.csv
@@ -1,32 +1,0 @@
-url
-https://finance.yahoo.com/news/more-than-200-sp-500-companies-scrubbed-diversity-and-equity-from-annual-reports-in-2025-143039186.html
-https://finance.yahoo.com/news/millennial-asked-personally-know-millennial-194533129.html
-https://finance.yahoo.com/news/most-humbling-thing-ve-ever-194500540.html
-https://finance.yahoo.com/news/apple-inc-aapl-pay-apple-191213724.html
-https://finance.yahoo.com/news/nvidia-vs-microsoft-stock-first-131500147.html
-https://finance.yahoo.com/news/4-things-investors-warren-buffett-110116006.html
-https://finance.yahoo.com/news/street-calls-week-094036318.html
-https://finance.yahoo.com/news/secret-super-stocks-invested-1k-230119586.html
-https://finance.yahoo.com/news/apple-aapl-stock-upgraded-hold-161143005.html
-https://finance.yahoo.com/news/apple-inc-aapl-people-tired-101059778.html
-https://finance.yahoo.com/news/want-move-data-between-apple-090000002.html
-https://finance.yahoo.com/news/elon-musk-thinks-tesla-become-082200237.html
-https://finance.yahoo.com/news/foxconn-second-quarter-revenue-rises-075413362.html
-https://finance.yahoo.com/news/apple-races-box-office-glory-040031137.html
-https://finance.yahoo.com/news/why-netflix-stock-jumped-11-232856176.html
-https://finance.yahoo.com/news/apple-aapl-beat-estimates-again-161002952.html
-https://finance.yahoo.com/news/clara-technologies-8-launches-sales-144739786.html
-https://finance.yahoo.com/news/biggest-climate-fund-approves-record-133155646.html
-https://finance.yahoo.com/news/modern-day-plague-infostealers-expose-131643509.html
-https://finance.yahoo.com/news/latest-news-cloud-ai-healthcare-123735436.html
-https://finance.yahoo.com/news/apples-china-iphone-sales-finally-091256165.html
-https://finance.yahoo.com/news/chinese-sales-foreign-phone-makers-090647988.html
-https://finance.yahoo.com/news/nvidia-track-become-most-valuable-085709500.html
-https://finance.yahoo.com/news/global-funds-position-wild-ride-080000131.html
-https://finance.yahoo.com/news/robinhood-launches-tokenized-us-stocks-064301025.html
-https://finance.yahoo.com/news/circus-se-gets-awarded-international-060300549.html
-https://finance.yahoo.com/news/most-popular-stocks-funds-investors-june-050100921.html
-https://finance.yahoo.com/news/consumer-electronics-stocks-q1-highlights-033425980.html
-https://finance.yahoo.com/news/veteran-fund-manager-makes-bold-205034775.html
-https://finance.yahoo.com/news/top-research-reports-apple-oracle-202500964.html
-https://finance.yahoo.com/news/nvidia-could-something-no-company-185118270.html
