--- news/temp/AMZN_temp.csv
@@ -1,29 +1,0 @@
-url
-https://finance.yahoo.com/news/more-than-200-sp-500-companies-scrubbed-diversity-and-equity-from-annual-reports-in-2025-143039186.html
-https://finance.yahoo.com/news/millennial-asked-personally-know-millennial-194533129.html
-https://finance.yahoo.com/news/sneak-vodka-pubs-save-drinks-130000243.html
-https://finance.yahoo.com/news/ceo-went-folding-clothes-nordstrom-120700537.html
-https://finance.yahoo.com/news/insider-alert-amazon-exec-offloads-231858602.html
-https://finance.yahoo.com/news/amazon-vs-microsoft-cloud-computing-093500337.html
-https://finance.yahoo.com/news/jefferies-survey-u-amazon-shoppers-093002163.html
-https://finance.yahoo.com/news/expect-markets-week-tariff-deadline-090000741.html
-https://finance.yahoo.com/news/mili-gears-amazon-prime-day-064100465.html
-https://finance.yahoo.com/news/secret-super-stocks-invested-1k-230119586.html
-https://finance.yahoo.com/news/amazon-doubles-down-robotaxis-amzn-160002772.html
-https://finance.yahoo.com/news/win-battle-u-retail-media-103502265.html
-https://finance.yahoo.com/news/did-hyperscale-clouds-q1-2025-103003101.html
-https://finance.yahoo.com/news/amazon-com-inc-amzn-jim-101203780.html
-https://finance.yahoo.com/news/retailers-cashing-trade-war-fp-100035026.html
-https://finance.yahoo.com/news/consumers-embrace-early-school-shopping-090000595.html
-https://finance.yahoo.com/news/elon-musk-thinks-tesla-become-082200237.html
-https://finance.yahoo.com/news/ultimate-growth-stock-buy-1-081000398.html
-https://finance.yahoo.com/news/apple-races-box-office-glory-040031137.html
-https://finance.yahoo.com/news/cargojet-buy-rating-target-confirmed-151657308.html
-https://finance.yahoo.com/news/t-optimizes-portfolio-strategic-divestiture-140600177.html
-https://finance.yahoo.com/news/confident-amazon-amzn-ability-deliver-122945177.html
-https://finance.yahoo.com/news/first-big-winners-race-create-110000917.html
-https://finance.yahoo.com/news/prediction-2-incredible-artificial-intelligence-095700671.html
-https://finance.yahoo.com/news/growth-stock-skyrocketed-225-000-085100019.html
-https://finance.yahoo.com/news/jefferies-maintains-buy-nike-amid-073149875.html
-https://finance.yahoo.com/news/billionaire-bill-ackman-may-next-071200937.html
-https://finance.yahoo.com/news/amazon-hits-1-million-robots-013112699.html
