--- news/temp/GOOG_temp.csv
@@ -1,7 +1,0 @@
-url
-https://finance.yahoo.com/news/most-humbling-thing-ve-ever-194500540.html
-https://finance.yahoo.com/news/why-wall-street-still-trusts-165913439.html
-https://finance.yahoo.com/news/latest-news-cloud-ai-healthcare-123735436.html
-https://finance.yahoo.com/news/cme-groups-quarterly-earnings-preview-110536788.html
-https://finance.yahoo.com/news/first-big-winners-race-create-110000917.html
-https://finance.yahoo.com/news/google-invests-nuclear-fusion-power-065339574.html
