--- news/temp/META_temp.csv
@@ -1,26 +1,0 @@
-url
-https://finance.yahoo.com/news/more-than-200-sp-500-companies-scrubbed-diversity-and-equity-from-annual-reports-in-2025-143039186.html
-https://finance.yahoo.com/news/some-complacency-has-crept-in-how-fomo-and-speculative-bets-are-driving-the-2025-market-rally-133011080.html
-https://finance.yahoo.com/news/threads-turns-two-continues-gain-203731702.html
-https://finance.yahoo.com/news/expect-magnificent-seven-second-half-090000904.html
-https://finance.yahoo.com/news/meta-eu-ad-troubles-could-170217733.html
-https://finance.yahoo.com/news/prediction-artificial-intelligence-ai-stock-134300513.html
-https://finance.yahoo.com/news/openai-betting-millions-building-ai-115400155.html
-https://finance.yahoo.com/news/3-magnificent-seven-stocks-crush-093000236.html
-https://finance.yahoo.com/news/best-growth-stocks-id-buy-083000688.html
-https://finance.yahoo.com/news/better-artificial-intelligence-stock-nvidia-082000291.html
-https://finance.yahoo.com/news/zuck-bucks-back-time-theyre-232735396.html
-https://finance.yahoo.com/news/whatsapp-business-just-got-smarter-214613074.html
-https://finance.yahoo.com/news/meta-platforms-stock-looks-cheap-174740548.html
-https://finance.yahoo.com/news/applovin-stock-surges-47-3-173300452.html
-https://finance.yahoo.com/news/meta-upgraded-pnc-financial-downgraded-134547568.html
-https://finance.yahoo.com/news/modern-day-plague-infostealers-expose-131643509.html
-https://finance.yahoo.com/news/eu-says-continue-rolling-ai-122350550.html
-https://finance.yahoo.com/news/first-big-winners-race-create-110000917.html
-https://finance.yahoo.com/news/jim-cramer-says-meta-ridiculously-065508136.html
-https://finance.yahoo.com/news/jim-cramer-says-nvidia-artificial-065506296.html
-https://finance.yahoo.com/news/jim-cramer-notes-microsoft-doesn-065504785.html
-https://finance.yahoo.com/news/meta-launches-superintelligence-labs-accelerate-065337958.html
-https://finance.yahoo.com/news/microsoft-unveils-mai-dxo-ai-065335220.html
-https://finance.yahoo.com/news/3-market-beating-stocks-competitive-043259559.html
-https://finance.yahoo.com/news/scale-ais-alexandr-wang-went-023055270.html
