--- news/temp/MSFT_temp.csv
@@ -1,29 +1,0 @@
-url
-https://finance.yahoo.com/news/the-real-risk-of-a-recession-and-other-key-points-about-the-us-economy-right-now-164822925.html
-https://finance.yahoo.com/news/nvidia-vs-microsoft-stock-first-131500147.html
-https://finance.yahoo.com/news/amazon-vs-microsoft-cloud-computing-093500337.html
-https://finance.yahoo.com/news/expect-magnificent-seven-second-half-090000904.html
-https://finance.yahoo.com/news/da-davidson-calls-microsoft-msft-171240998.html
-https://finance.yahoo.com/news/2-unstoppable-growth-stocks-hold-151000255.html
-https://finance.yahoo.com/news/prediction-artificial-intelligence-ai-stock-134300513.html
-https://finance.yahoo.com/news/openai-partnership-microsoft-good-says-120002534.html
-https://finance.yahoo.com/news/did-hyperscale-clouds-q1-2025-103003101.html
-https://finance.yahoo.com/news/vistra-corp-vst-experiencing-trend-101935121.html
-https://finance.yahoo.com/news/billionaire-bill-gates-66-foundations-095200608.html
-https://finance.yahoo.com/news/3-magnificent-seven-stocks-crush-093000236.html
-https://finance.yahoo.com/news/elon-musk-thinks-tesla-become-082200237.html
-https://finance.yahoo.com/news/microsoft-closing-local-operations-pakistan-174213766.html
-https://finance.yahoo.com/news/lunit-partners-microsoft-develop-ai-171529744.html
-https://finance.yahoo.com/news/spot-outliers-hyperscaler-microsoft-early-165514646.html
-https://finance.yahoo.com/news/gitlabs-devsecops-growth-momentum-continues-142500489.html
-https://finance.yahoo.com/news/ibms-neudesic-boosts-cloud-security-141100749.html
-https://finance.yahoo.com/news/investors-heavily-search-microsoft-corporation-130003666.html
-https://finance.yahoo.com/news/latest-news-cloud-ai-healthcare-123735436.html
-https://finance.yahoo.com/news/prediction-2-incredible-artificial-intelligence-095700671.html
-https://finance.yahoo.com/news/2-dividend-stocks-could-help-090500761.html
-https://finance.yahoo.com/news/jim-cramer-says-meta-ridiculously-065508136.html
-https://finance.yahoo.com/news/jim-cramer-says-nvidia-artificial-065506296.html
-https://finance.yahoo.com/news/jim-cramer-notes-microsoft-doesn-065504785.html
-https://finance.yahoo.com/news/microsoft-unveils-mai-dxo-ai-065335220.html
-https://finance.yahoo.com/news/robinhood-launches-tokenized-us-stocks-064301025.html
-https://finance.yahoo.com/news/1-profitable-stock-target-week-044028055.html
