--- news/temp/NFLX_temp.csv
@@ -1,3 +1,0 @@
-url
-https://finance.yahoo.com/news/another-company-scrapping-unlimited-time-200021992.html
-https://finance.yahoo.com/news/buy-netflix-stock-july-17-070200248.html
