--- news/temp/TSLA_temp.csv
@@ -1,40 +1,0 @@
-url
-https://finance.yahoo.com/news/south-koreas-lg-energy-solution-235208511.html
-https://finance.yahoo.com/news/trump-derides-musk-third-political-232001666.html
-https://finance.yahoo.com/news/elon-musks-america-party-reportedly-194932726.html
-https://finance.yahoo.com/news/most-humbling-thing-ve-ever-194500540.html
-https://finance.yahoo.com/news/musk-worry-companies-not-politics-194245810.html
-https://finance.yahoo.com/news/costcos-ev-charging-developed-spacex-173102383.html
-https://finance.yahoo.com/news/tariff-dealine-fed-minutes-other-170002333.html
-https://finance.yahoo.com/news/veteran-tesla-analyst-drops-4-163400015.html
-https://finance.yahoo.com/news/cathie-wood-just-went-bargain-142500090.html
-https://finance.yahoo.com/news/weve-got-act-together-warns-230128157.html
-https://finance.yahoo.com/news/secret-super-stocks-invested-1k-230119586.html
-https://finance.yahoo.com/news/where-tesla-stock-3-years-223000525.html
-https://finance.yahoo.com/news/elon-musks-father-says-warned-170019941.html
-https://finance.yahoo.com/news/models-robotaxi-ai-why-tesla-160344388.html
-https://finance.yahoo.com/news/amazon-doubles-down-robotaxis-amzn-160002772.html
-https://finance.yahoo.com/news/why-everyone-talking-soundhound-ai-130300212.html
-https://finance.yahoo.com/news/4-ways-musk-trump-clash-120607099.html
-https://finance.yahoo.com/news/weekend-reeves-trump-spending-bill-welfare-101222842.html
-https://finance.yahoo.com/news/tesla-inc-tsla-people-gleeful-100736439.html
-https://finance.yahoo.com/news/ghost-factories-warning-sign-green-100025377.html
-https://finance.yahoo.com/news/3-magnificent-seven-stocks-crush-093000236.html
-https://finance.yahoo.com/news/elon-musk-thinks-tesla-become-082200237.html
-https://finance.yahoo.com/news/elon-musks-71-billion-wealth-033053767.html
-https://finance.yahoo.com/news/elon-musk-backs-warren-buffetts-230021114.html
-https://finance.yahoo.com/news/mark-cuban-elon-musk-something-213106722.html
-https://finance.yahoo.com/news/behold-epic-flop-cybertruck-140025865.html
-https://finance.yahoo.com/news/tesla-sales-down-13-5-171137994.html
-https://finance.yahoo.com/news/agl-acquires-south-australia-virtual-154111999.html
-https://finance.yahoo.com/news/gm-beats-tesla-ev-growth-141600228.html
-https://finance.yahoo.com/news/navitas-gan-power-ics-gaining-140200588.html
-https://finance.yahoo.com/news/teslas-worst-ever-q2-vehicle-140000793.html
-https://finance.yahoo.com/news/tesla-very-strong-june-looks-132511850.html
-https://finance.yahoo.com/news/expert-predicted-tesla-stock-crash-130124915.html
-https://finance.yahoo.com/news/stocks-making-big-moves-week-130037835.html
-https://finance.yahoo.com/news/uk-best-selling-cars-revealed-112804683.html
-https://finance.yahoo.com/news/latest-news-electric-vehicles-revolutionizing-103934816.html
-https://finance.yahoo.com/news/jim-cramer-tesla-total-dice-065449016.html
-https://finance.yahoo.com/news/most-popular-stocks-funds-investors-june-050100921.html
-https://finance.yahoo.com/news/elon-finally-woken-musk-battles-040027959.html
